+++ changelog.md
@@ -13,6 +13,7 @@
 
 
 * [Changelog](#changelog)
+  * [Releases v1.2.0](#releases-v120)
   * [Releases v1.1.2](#releases-v112)
   * [Releases v1.1.1](#releases-v111)
  
@@ -20,6 +21,12 @@
 ---
 
 ## Changelog
+
+### Releases v1.2.0
+
+1. Auto detect ESP32 core and use either built-in LittleFS or [LITTLEFS](https://github.com/lorol/LITTLEFS) library
+2. Update `library.json` to use new `headers` for PIO
+3. Update `platformio.ini`
 
 ### Releases v1.1.2